Split out worker protocol template definitions from declarations
This is generally good practice, and will become more important in a
moment.
