+++ src/libstore/remote-store.cc
@@ -6,6 +6,7 @@
 #include "build-result.hh"
 #include "remote-store.hh"
 #include "worker-protocol.hh"
+#include "worker-protocol-impl.hh"
 #include "archive.hh"
 #include "globals.hh"
 #include "derivations.hh"
